+++ examples/basic/src/App.tsx
@@ -77,7 +77,7 @@ function App() {
         <EditorEventListener
           on={(event) => {
             if (event.type === 'mutation') {
-              setValue(event.snapshot)
+              setValue(event.value)
             }
           }}
         />