+++ packages/editor/src/behaviors/behavior.links.ts
@@ -1,9 +1,10 @@
 import type {EditorSchema} from '../editor/define-schema'
 import * as selectors from '../selectors'
+import {looksLikeUrl} from '../utils/looks-like-url'
 import {defineBehavior} from './behavior.types'
 
 /**
- * @alpha
+ * @beta
  */
 export type LinkBehaviorsConfig = {
   linkAnnotation?: (context: {
@@ -13,7 +14,7 @@ export type LinkBehaviorsConfig = {
 }
 
 /**
- * @alpha
+ * @beta
  */
 export function createLinkBehaviors(config: LinkBehaviorsConfig) {
   const pasteLinkOnSelection = defineBehavior({
@@ -79,13 +80,4 @@ export function createLinkBehaviors(config: LinkBehaviorsConfig) {
   const linkBehaviors = [pasteLinkOnSelection, pasteLinkAtCaret]
 
   return linkBehaviors
-}
-
-function looksLikeUrl(text: string) {
-  let looksLikeUrl = false
-  try {
-    new URL(text)
-    looksLikeUrl = true
-  } catch {}
-  return looksLikeUrl
 }