+++ packages/editor/src/editor/create-editor.ts
@@ -24,13 +24,18 @@ import {
   type EditorEmittedEvent,
   type InternalEditorEvent,
 } from './editor-machine'
+import {getEditorSnapshot} from './editor-selector'
+import type {EditorSnapshot} from './editor-snapshot'
 import {defaultKeyGenerator} from './key-generator'
 import {createEditableAPI} from './plugins/createWithEditableAPI'
 
 /**
- * @alpha
+ * @public
  */
 export type EditorConfig = {
+  /**
+   * @beta
+   */
   behaviors?: Array<Behavior>
   keyGenerator?: () => string
   /**
@@ -51,7 +56,7 @@ export type EditorConfig = {
 )
 
 /**
- * @alpha
+ * @public
  */
 export type EditorEvent = PickFromUnion<
   InternalEditorEvent,
@@ -75,9 +80,10 @@ export type EditorEvent = PickFromUnion<
 >
 
 /**
- * @alpha
+ * @public
  */
 export type Editor = {
+  getSnapshot: () => EditorSnapshot
   send: (event: EditorEvent) => void
   on: ActorRef<Snapshot<unknown>, EventObject, EditorEmittedEvent>['on']
   _internal: {
@@ -126,6 +132,11 @@ function createEditorFromActor(editorActor: EditorActor): Editor {
   const editable = createEditableAPI(slateEditor.instance, editorActor)
 
   return {
+    getSnapshot: () =>
+      getEditorSnapshot({
+        editorActorSnapshot: editorActor.getSnapshot(),
+        slateEditorInstance: slateEditor.instance,
+      }),
     send: (event) => {
       editorActor.send(event)
     },